+++ choco-solver/src/main/java/org/chocosolver/solver/constraints/Propagator.java
@@ -93,30 +93,84 @@ import static org.chocosolver.solver.variables.events.PropagatorEventType.CUSTOM
  */
 public abstract class Propagator<V extends Variable> implements Serializable, ICause, Identity, Comparable<Propagator> {
 
-    //***********************************************************************************
-    // VARIABLES
-    //***********************************************************************************
-
+    /**
+     * For serialization purpose.
+     */
     private static final long serialVersionUID = 2L;
-    protected static final short NEW = 0, REIFIED = 1, ACTIVE = 2, PASSIVE = 3;
 
-    // propagator attributes
-    private final int ID; // unique id of this
-    private short state;  // 0 : new -- 1 : active -- 2 : passive
-    private Operation[] operations; // propagator state operations
-    private int nbPendingEvt = 0;   // counter of enqued records -- usable as trigger for complex algorithm
+    /**
+     * Status of this propagator on creation.
+     */
+    protected static final short NEW = 0;
+
+    /**
+     * Status of this propagagator when reified.
+     */
+    protected static final short REIFIED = 1;
+
+    /**
+     * Status of the propagator when activated (ie, after initial propagation).
+     */
+    protected static final short ACTIVE = 2;
+
+    /**
+     * Status of the propagator when entailed.
+     */
+    protected static final short PASSIVE = 3;
+
+    /**
+     * Unique ID of this propagator.
+     */
+    private final int ID;
+
+    /**
+     * Current status of this propagator.
+     * In: {@link #NEW}, {@link #REIFIED}, {@link #ACTIVE} and {@link #PASSIVE}.
+     */
+    private short state;
+
+    /**
+     * Backtrackable operations to maintain the status on backtrack.
+     */
+    private Operation[] operations;
+
+    /**
+     * Counter of events to be propagated by this propagator.
+     */
+    private int nbPendingEvt = 0;
+
+    /**
+     * Priority of this propagator.
+     * Mix between arity and compexity.
+     */
     protected final PropagatorPriority priority;
-    protected final boolean reactToFineEvt;
-    // references
-    protected Constraint constraint; // declaring constraint
-    protected final Solver solver;   // solver of this propagator
-    // variable related information
-    protected V[] vars;// List of <code>variable</code> objects -- a variable can occur more than once, but it could not have the same index
-    private int[] vindices;// index of this within the list of propagator of the i^th variable
 
-    //***********************************************************************************
-    // CONSTRUCTORS
-    //***********************************************************************************
+    /**
+     * Set to <tt>true</tt> to indidates that this propagator reacts to fine event.
+     * If set to <tt>false</tt>, the method {@link #propagate(int, int)} will never be called.
+     */
+    protected final boolean reactToFineEvt;
+
+    /**
+     * Encapsuling constraint.
+     */
+    protected Constraint constraint;
+
+    /**
+     * Reference to the solver declaring this propagator.
+     */
+    protected final Solver solver;
+
+    /**
+     * List of variables this propagators deal with.
+     * A variable can occur more than once, but it is considered then as n distinct variables.
+     */
+    protected V[] vars;
+
+    /**
+     * Index of this propagator within each variable's propagators.
+     */
+    private int[] vindices;
 
     /**
      * Creates a new propagator to filter the domains of vars.
@@ -204,13 +258,12 @@ public abstract class Propagator<V extends Variable> implements Serializable, IC
         if (solver.getEngine() != NoPropagationEngine.SINGLETON && solver.getEngine().isInitialized()) {
             solver.getEngine().updateInvolvedVariables(this);
         }
-        if (isActive()) {
-            for (int v = tmp.length; v < vars.length; v++) {
-                vars[v].recordMask(getPropagationConditions(v));
-            }
-        }
     }
 
+    /**
+     * Creates links between this propagator and its variables.
+     * The propagator will then be referenced in each of its variables.
+     */
     protected void linkVariables() {
         for (int v = 0; v < vars.length; v++) {
             vindices[v] = vars[v].link(this, v);
@@ -311,6 +364,7 @@ public abstract class Propagator<V extends Variable> implements Serializable, IC
      * Add the coarse event recorder into the engine
      *
      * @param evt event type
+     * @throws ContradictionException if the propagation encounters inconsistency.
      */
     public final void forcePropagate(PropagatorEventType evt) throws ContradictionException {
         solver.getEngine().delayedPropagation(this, evt);
@@ -323,10 +377,6 @@ public abstract class Propagator<V extends Variable> implements Serializable, IC
         assert isStateLess() : "the propagator is already active, it cannot set active";
         state = ACTIVE;
         solver.getEnvironment().save(operations[NEW]);
-        // update activity mask of variables
-        for (int v = 0; v < vars.length; v++) {
-            vars[v].recordMask(getPropagationConditions(v));
-        }
     }
 
     /**
@@ -336,10 +386,6 @@ public abstract class Propagator<V extends Variable> implements Serializable, IC
         assert isReifiedAndSilent() : "the propagator was not in a silent reified state";
         state = ACTIVE;
         solver.getEnvironment().save(operations[REIFIED]);
-        // update activity mask of variables
-        for (int v = 0; v < vars.length; v++) {
-            vars[v].recordMask(getPropagationConditions(v));
-        }
     }
 
     /**
@@ -425,6 +471,13 @@ public abstract class Propagator<V extends Variable> implements Serializable, IC
         return arity;
     }
 
+    /**
+     * Return the dynamic priority of this propagator.
+     * It excludes from the arity variables instantiated.
+     * But may be time consuming.
+     * @return a more accurate priority excluding instantiated variables.
+     */
+    @SuppressWarnings("unused")
     public int dynPriority() {
         int arity = 0;
         for (int i = 0; i < vars.length && arity <= 3; i++) {
@@ -510,6 +563,7 @@ public abstract class Propagator<V extends Variable> implements Serializable, IC
     /**
      * @return the index of the propagator within its variables
      */
+    @SuppressWarnings("unused")
     public int[] getVIndices() {
         return vindices;
     }
@@ -521,6 +575,7 @@ public abstract class Propagator<V extends Variable> implements Serializable, IC
      * @param idx old index
      * @param val new index
      */
+    @SuppressWarnings("unused")
     public void setVIndices(int idx, int val) {
         vindices[idx] = val;
     }