+++ choco-solver/src/test/java/org/chocosolver/solver/constraints/DynamicPostTest.java
@@ -1,21 +1,21 @@
 /**
  * Copyright (c) 2015, Ecole des Mines de Nantes
  * All rights reserved.
- *
+ * <p>
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
+ * notice, this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
+ * notice, this list of conditions and the following disclaimer in the
+ * documentation and/or other materials provided with the distribution.
  * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *    This product includes software developed by the <organization>.
+ * must display the following acknowledgement:
+ * This product includes software developed by the <organization>.
  * 4. Neither the name of the <organization> nor the
- *    names of its contributors may be used to endorse or promote products
- *    derived from this software without specific prior written permission.
- *
+ * names of its contributors may be used to endorse or promote products
+ * derived from this software without specific prior written permission.
+ * <p>
  * THIS SOFTWARE IS PROVIDED BY <COPYRIGHT HOLDER> ''AS IS'' AND ANY
  * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -41,7 +41,7 @@ import org.chocosolver.solver.variables.IntVar;
 import org.chocosolver.solver.variables.VF;
 import org.chocosolver.solver.variables.VariableFactory;
 import org.chocosolver.util.ESat;
-import org.chocosolver.util.tools.StringUtils;
+import org.chocosolver.util.ProblemMaker;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
@@ -209,7 +209,7 @@ public class DynamicPostTest {
                 LCF.and(aSBetter, bBetter),
                 LCF.and(aBetter, bSBetter));
         // END extra variables/constraints for guided improvement algorithm
-        solver.set(ISF.lexico_LB(a,b,c,lbA,lbB));
+        solver.set(ISF.lexico_LB(a, b, c, lbA, lbB));
         int nbSolution = 0;
         while (solver.findSolution()) {
             int bestA;
@@ -267,48 +267,29 @@ public class DynamicPostTest {
         solver.unpost(c);
     }
 
-	@Test(groups = "1s")
-	public void testCostas() {
-		Solver s1 = costasArray(7,false);
-		Solver s2 = costasArray(7,true);
+    @Test(groups = "1s")
+    public void testCostas() {
+        Solver s1 = costasArray(7, false);
+        Solver s2 = costasArray(7, true);
 
-		s1.findAllSolutions();
-		System.out.println(s1.getMeasures().getSolutionCount());
+        s1.findAllSolutions();
+        System.out.println(s1.getMeasures().getSolutionCount());
 
-		s2.findAllSolutions();
+        s2.findAllSolutions();
 
-		System.out.println(s2.getMeasures().getSolutionCount());
-		Assert.assertEquals(s1.getMeasures().getSolutionCount(),s2.getMeasures().getSolutionCount());
-	}
+        System.out.println(s2.getMeasures().getSolutionCount());
+        Assert.assertEquals(s1.getMeasures().getSolutionCount(), s2.getMeasures().getSolutionCount());
+    }
 
-	private Solver costasArray(int n, boolean dynamic){
-		Solver solver = new Solver("CostasArrays");
-		IntVar[] vars, vectors;
-		vars = VariableFactory.enumeratedArray("v", n, 0, n - 1, solver);
-		vectors = new IntVar[(n*(n-1))/2];
-		IntVar[][] diff = new IntVar[n][n];
-		int idx = 0;
-		for (int i = 0; i < n; i++) {
-			for (int j = i+1; j < n; j++) {
-				IntVar k = VariableFactory.enumerated(StringUtils.randomName(), -n, n, solver);
-				solver.post(ICF.arithm(k,"!=",0));
-				solver.post(IntConstraintFactory.sum(new IntVar[]{vars[i],k},vars[j]));
-				vectors[idx] = VariableFactory.offset(k, 2 * n * (j - i));
-				diff[i][j] = k;
-				idx++;
-			}
-		}
-		solver.post(ICF.alldifferent(vars, "AC"));
-		solver.post(ICF.alldifferent(vectors, "BC"));
-		// symmetry-breaking
-		solver.post(ICF.arithm(vars[0],"<",vars[n-1]));
-//		SMF.limitTime(solver, 20000);
-		solver.set(ISF.domOverWDeg(vectors,0));
+    private Solver costasArray(int n, boolean dynamic) {
+        Solver solver = ProblemMaker.makeCostasArrays(n);
+        IntVar[] vectors = (IntVar[]) solver.getHook("vectors");
+        solver.set(ISF.domOverWDeg(vectors, 0));
 
-		if(dynamic){
-			// should not change anything (the constraint is already posted)
-			solver.plugMonitor((IMonitorSolution) () -> solver.post(ICF.alldifferent(vectors,"BC")));
-		}
-		return solver;
-	}
+        if (dynamic) {
+            // should not change anything (the constraint is already posted)
+            solver.plugMonitor((IMonitorSolution) () -> solver.post(ICF.alldifferent(vectors, "BC")));
+        }
+        return solver;
+    }
 }